mqtt-to-db-sensor-data/requirements.txt
paho-mqtt
dotenv
